extracts the dark mode functionality from app.jsx
