+++ src/App.jsx
@@ -1,63 +1,20 @@
-import React, { useState, useMemo, createContext } from "react";
+import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Header from "./components/Header";
 import Departments from "./components/Departments";
 import Events from "./components/Events";
 import About from "./components/About";
-import { ThemeProvider, createTheme } from "@mui/material/styles";
-import { CssBaseline } from "@mui/material";
-import IconButton from "@mui/material/IconButton";
-import Brightness4Icon from "@mui/icons-material/Brightness4";
-import Brightness7Icon from "@mui/icons-material/Brightness7";
-
-const ColorModeContext = createContext({ toggleColorMode: () => {} });
 
 const App = () => {
-  const [mode, setMode] = useState("light");
-  const colorMode = useMemo(
-    () => ({
-      toggleColorMode: () => {
-        setMode((prevMode) => (prevMode === "light" ? "dark" : "light"));
-      },
-    }),
-    []
-  );
-
-  const theme = React.useMemo(
-    () =>
-      createTheme({
-        palette: {
-          mode,
-        },
-      }),
-    [mode]
-  );
-
   return (
-    <ColorModeContext.Provider value={colorMode}>
-      <ThemeProvider theme={theme}>
-        <CssBaseline />
-        <BrowserRouter>
-          <Header />
-          <Routes>
-            <Route exact path="/" element={<Events />} />
-            <Route path="/departments" element={<Departments />} />
-            <Route path="/about" element={<About />} />
-          </Routes>
-          <IconButton
-            sx={{ ml: 1 }}
-            onClick={colorMode.toggleColorMode}
-            color="inherit"
-          >
-            {theme.palette.mode === "dark" ? (
-              <Brightness7Icon />
-            ) : (
-              <Brightness4Icon />
-            )}
-          </IconButton>
-        </BrowserRouter>
-      </ThemeProvider>
-    </ColorModeContext.Provider>
+    <BrowserRouter>
+      <Header />
+      <Routes>
+        <Route exact path="/" element={<Events />} />
+        <Route path="/departments" element={<Departments />} />
+        <Route path="/about" element={<About />} />
+      </Routes>
+    </BrowserRouter>
   );
 };
